Update to ddwrt.py - Using LAN instead of Wireless
Current code only checks for wireless clients connected to this router. This is not effective when using multiple routers or additional access points as they appear in the LAN list, not the wireless.
I noticed this when I moved from asuswrt to ddwrt. Modified ddwrt to use LAN active client list instead of only Wireless active clients.